requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
backports.functools-lru-cache==2.0.0
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
cachetools==5.5.2
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contractions==0.1.73
fastapi==0.115.11
filelock==3.17.0
fsspec==2025.3.0
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-play-scraper==1.2.7
googleapis-common-protos==1.69.1
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
hyperframe==5.2.0
idna==2.10
indic-transliteration==2.3.69
Jinja2==3.1.6
joblib==1.4.2
langdetect==1.0.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numexpr @ file:///C:/b/abs_05o8p7bfml/croot/numexpr_1730215959182/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp311-cp311-win_amd64.whl#sha256=46eb5850478cda64bda6d2f2d3d6ad8ceeb7cc168b83532e4f9dbbcb42e22ab2
ollama==0.4.7
packaging==24.2
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp311-cp311-win_amd64.whl#sha256=3a9ed5b5006a5fad640788cd8385485f98b18ac6ff9294cb25b18f3931fed352
proto-plus==1.26.0
protobuf==5.29.3
pyahocorasick==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
roman==5.0
rsa==4.9
safetensors==0.5.3
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
textsearch==0.0.24
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
